requirements.txt
adafruit-ampy==1.1.0
bitarray==3.8.0
bitstring==4.3.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.4
esptool==5.1.0
gitdb==4.0.12
GitPython==3.1.46
intelhex==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
mpremote==1.27.0
pathlib==1.0.1
platformdirs==4.5.1
pycparser==3.0
Pygments==2.19.2
pyserial==3.5
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
reedsolo==1.7.0
rich==14.3.1
rich-click==1.9.6
shiboken6==6.10.1
smmap==5.0.2
